+++ MemoryMap/MemoryMap.hpp
@@ -9,8 +9,8 @@ public:
 	virtual bool CanRead(uint32_t address) const = 0;
 	virtual bool CanWrite(uint32_t address) const = 0;
 
-	virtual bool Read(uint32_t address, uint8_t& value) = 0;
-	virtual bool Write(uint32_t address, uint8_t value) = 0;
+	virtual bool Read(uint32_t address, uint32_t& value) = 0;
+	virtual bool Write(uint32_t address, uint32_t value) = 0;
 };
